requirements.txt
pandas==0.25.1
matplotlib==3.1.1
numpy==1.17.2
pylablib==1.1.0
ThorlabsPM100==1.2.2
python-usbtmc==0.8